refactor: extract terminal printing of Diagnosic
