+++ cjkfmt-cli/src/cli/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod args;
 pub(crate) mod check;
 pub(crate) mod format;
+pub(crate) mod utils;
